pyproject.toml
[project]
name = "proj-legal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "langchain-core>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
